Change to unified stream with separate cursors
Previously input and output was separated to their own boxes,
with input always at the bottom of the screen, and output did
not mess up with input in any way.

However, the same can be achieved by using separate cursors
for input and output, so that unexpected output will not
interfere with input.

At the same separate input buffer does not need to be kept
for editing, but the whole stream can be made editable.